refactor button/header dialog components to dialog
